Rcode File Upload
This is the first upload to the MetLip front end automation repository. These scripts are rough drafts for the meta data generator and plate data generator. The meta data generator is offered in a function layout as well. The meta data generator works just fine. Whereas, the plate data generator needs work in order to achieve flexibility to accommodate various meta data files.