Upload Sequence_generator_annotated
This script is the first version of the sequence generator which depends on two data frames. The first is Plate_Metadata which we generate using our developed RShiny App. The second is Acquired_Samples_Data which I produce manually using TAS project submission info. A future project will be to generate Acquired_Samples_Data automatically and essentially integrate with Plate_Metadata. The current version of Sequence_generator, uploaded here, relies on the use of for loops. The next step will be to transition to the use of a function. That way we can eventually develop this as a ShinyApp.